+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+Minicrypto 0.3.0 - 29 Jul, 2026
+-------------------------------
+- `minicrypto.argon2`.
+
+Minicrypto 0.2.0 - 29 Jul, 2026
+-------------------------------
+- `minicrypto.blake2b`
+- All functions are now `@nogc`.
+
 Minicrypto 0.1.0 - 28 Jul, 2026
 -------------------------------
 Initial release.